// src/app/login/utils/login.interface.ts
export interface LoginCrendential {
    user?: string | null | undefined;
    password?: string | null | undefined;
    application: string;
    plant?: string;
    googleImage?: string;
}

export interface GooglePayload {
    iss: string;
    nbf: number;
    aud: string;
    sub: string;
    hd: string;
    email: string;
    email_verified: boolean;
    azp: string;
    name: string;
    picture: string;
    given_name: string;
    family_name: string;
    iat: number;
    exp: number;
    jti: string;
}